Add player class and player_colors
Add the Player class, which represents players of the game.
This class will be owned by ScorchedModel.

Add player_colors to colors.xml. I couldn't figure out how to store an array
of colors in the resource fork, so I stored the colors as strings instead.
Perhaps later they can be ints or something.