First shot at moving tank turrets.
My first attempt at having some kind of game logic going on.
My initial instinct was to serialize everything into the main() thread.
I did this because I believed that holding down a key would not give multiple
onKeyDown() calls. As it turns out, this is incorrect.

This code can be simplified a lot. A lot of code will be moved into
onKeyDown().