Fix drawScreen() display bug
There was a bug where we were beginning our drawing routine on a slot which
was a multiple of 3 rather than 2. This resulted in inconsistent appearance
when the user panned the screen, since a different set of slots would be
selected for the bezier curve. Also, it was sub-optimal because we were
drawing more than we needed to draw.