Make viewX = mZoom * gameX rather than the reverse
Change the coordinate system so that we have
viewX = mZoom * gameX rather than
gameX = mZoom * viewX.

In other words, mZoom is now the reciprocal of what it used to be. This
eliminates an expensive division to get slotWidth
(formerly slotWidth = 1/mZoom, now slotWidth = mZoom).
It also tends to minimize numerical errors by keeping mZoom in a nicer range
(0.5 - 60 or so) than before.

gameXtoOnscreenX() tends to happen a lot more than onscreenXtoGameX().
multiply is cheaper than divide. Therefore, it's good that we now have
gameXtoOnscreenX implemented as a multiply rather than a divide.

De-optimize zoomIn and zoomOut because I'm too lazy to fix the optimized
version. These functions only happen when the user requests them, and then
only as "single shots," so they're not worth optimizing anyway.