Rework GameState stuff
Create a better framework for game state. States are now first-class objects
which receive callbacks at strategic times. Also drew an ASCII-art diagram of
how the class states are laid out.

Fixed some comments about "blocking in the GUI thread"-- technically, blocking
in the GUI thread is allowed. Blocking for long periods of time is bad, of
course.

Use canvas.drawColor to clear the Canvas-- it is the best way.