Refactor the pan / zoom and GameState code
Refactor GameState code so that all user input flows through GameState, rather
than just some, as before. This means no user-triggered pan / zoom in states
other than the human player state, for now. Treat ZoomIn and ZoomOut as normal
buttons to simplify the code.

Kill Graphics.mNeedRedraw. The GameState should keep track of whether or not
the screen needs to be redrawn. (Only the GameState knows if it is animating
something, for example.)

Move the code which alters ViewSettings into the ViewSettings class. Formerly,
it was all over Graphics, which is not ideal-- since it really is a
ViewSettings thing. Now, we do have to pass around screen height and width,
though.  Make the APIs mutation-oriented because that's how we roll on
Android. Centralize constraint-enforcement code in mutate().

Some other small refactorings to kill warnings or eliminate unecessary
statics.