Create ModelFactory
Create ModelFactory class, which represents the game settings we are
configuring in the NewGame activity.

Add a comment to AndroidManifest.xml about a possible improvement to
ReadStuffActivity (using @android:style/Theme.Dialog on it)

Figure out how to get callbacks from the ListView in GameSetupActivity.
Paste in a code example called "SeparatedListAdapter" which we may or may not
end up using.

Eliminate playerColors in XML, because it's just too annoying. Make an enum in
the Player class to deal with all of this. This makes serialization a lot
easier.