Comment out old stuff and add degree symbol
Comment out the unused player_colors array in values/colors.xml. Don't delete
it completely, because I might need to make another one of these in the
future, and documentation is hard to come by on the XML stuff.

Add a real degree symbol and better font size to the degree readout.