No need for internal synch in ModelFactory
We don't need for internal synch in ModelFactory, because the View tree is
single-threaded. Callbacks run one by one in the UI thread, never in parallel.
So there's no issues with callbacks stepping on each other's toes.

The internal synchronization that I had there earlier was half-hearted and not
completely correct for the unsynchronized case, anyway.