Start handling FIRE button
Start handling FIRE button. It's not fully implemented yet, but this is a step
in the right direction.

Create showArmory and hideArmory in preparation for implementing the power
selector.

Rename R.drawable.transparent_grey to R.drawable.armory_bg_color, to make it
obvious what the color is used for.

Get rid of unused mWeapSelLeftButton