Introduce EARTHMOVERs + fix collision damage
Introduce EARTHMOVER weapons which do no explosion damage and have a different
color explosion.

Fix some issues in Explosion.doDirectDamage.
There are two important radii: the bullseye radius and the safety radius.
Every projectile whose center is inside of the bullseye radius counts as a
direct hit which does full damage.
Projectiles which explode outside the bullseye radius deal damage
linearly proportional to how close they are to the safety radius. Projectiles
whose centers lie outside of the safety radius do no damage to a player.

Introduce a new function, Util.linearInterpolation, to do exactly that.
Linear interpolation is trivial in theory, but in practice the code looks
messy sometimes if it's not encapsulated in a function.

Reformat the way WeaponType objects are declared a little bit, to make it more
readable.