Reorganize WeaponType + create cluster bombs
Refactor WeaponType so that explosion attributes are in a separate type, which
not all weapons have. This avoids the proliferation of explosion attributes
creeping into the general type, which had to be zeroed for non-explosive
weapons.

Allow BallisticsState to track multiple projectiles and explosions at once,
rather than just one. Separate out the concept of detonation from the concept
of explosion. Not all weapons detonate explosively-- some create other weapons
instead.

Introduce the idea of hidden WeaponType objects. These are just an
implementation detail and can't be directly fired or bought by users.

Change Explosion.mX and mY to ints.