First cut at BuyWeaponsState
A first try at implementing the weapons-buying state. It seems like the
easiest way to do this will be to create a modal dialog box inside RunGameAct.

We again need to use a Runnable object to start the dialog from the GUI
thread.

The dialog box doesn't cover the entire screen, although I think I can get it
to cover most of it. I want the background to be the sky of the upcoming
round, since that will look nice. Another possibility I am considering is
setting the background to the color of the player who is buying weapons-- just
to make it clearer who is buying in hot-seat games.

Remove crusty old TODO comment in drawScreen() that has already been done.
Remove another obsolete old todo calling for a "floating arrow."