Rework the Armory in terms of arrays
Rework the WeaponType.Armory object in terms of arrays. The overhead of using
a TreeMap is higher than using an array, what with GC and all. We aren't
expecting a large number of elements, either (the number should be around a
dozen or so), so an array of ints won't break the bank.