Create new Cosmos object for cross-round state
Create a new object called Cosmos to represent cross-round state. We already
kind of have cross-round state in the form of the ModelFactory. However, we
need something that can store earnings for the players and armories for the
players.

Also split PlayerColor into its own file, PlayerColor.java. It doesn't really
need to be in Player.java.