Beef up HardBrain and fix up other brain stuff
Beef up HardBrain so that it really is harder than MediumBrain.
How is it better? It tries three initial fixes before deciding on one. It
refines 3x as much as MediumBrain. And its shots never get worse (it never
refines its shots so that the error is greater.)

Fix a flaw in our optimization strategy. Previously, we were allowing the
angle or the power to change, but not both. This could result in us getting
stuck forever on a local hill in the optimization space. Sometimes it is
possible to improve a shot only by tweaking both the power and the angle.